requirements.txt
python-gitlab>=6.4.0
requests>=2.32.0
slack-sdk>=3.23.0
